Added conversion to Irm, Changed to int16_t

+++ Sources/DOR.h
@@ -32,7 +32,7 @@ typedef struct AnalogThreadData
   OS_ECB* semaphore;
   uint8_t channelNb;
   float sample;
-  float samples[16];
+  int16_t samples[16];
 } TAnalogThreadData;
 
 bool DOR_Init(const TDORSetup* const dorSetup);
